WP-CLIを導入する #1
resolved #1

+++ build.sh
@@ -1,4 +1,9 @@
-curl -O https://ja.wordpress.org/latest-ja.zip
-unzip latest-ja.zip
-rm latest-ja.zip
+#!/bin/sh
+
+# WordPressの展開
+if [ ! -e ./wordpress ]; then
+  curl -O https://ja.wordpress.org/latest-ja.zip
+  unzip latest-ja.zip
+  rm latest-ja.zip
+fi
 docker-compose up
